TimeZoneLookup

Base.delete!

Base.insert!

DataStructures.insert_node!

Base.delete! — Method

delete!(tree::RBTree, key, p, tol)

Deletes key from tree, if present, else returns the unmodified tree.

Base.insert! — Method

insert!(tree, key, p, tol = 1e-15)

Inserts key in the tree if it is not present.

DataStructures.insert_node! — Function

insert_node!(tree::RBTree, node::RBTreeNode, p::V, tol = 1e-15)

Inserts node at proper location by traversing through the tree in a binary-search-tree fashion.